team21_fcs_project/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==3.2.8
django-countries==7.2.1
django-crispy-forms==1.13.0
django-db-logger==0.1.10
django-formtools==2.3
django-otp==1.1.1
django-phonenumber-field==5.2.0
django-social-share==2.2.1
django-two-factor-auth==1.13.1
idna==3.3
phonenumbers==8.12.36
Pillow==8.3.2
pycodestyle==2.8.0
pyotp==2.6.0
python-magic==0.4.24
pytz==2021.3
qrcode==6.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
